Split Zarr reducer sinks
