+++ src/refiner/pipeline/sinks/reducer/__init__.py
@@ -1,7 +1,13 @@
 from refiner.pipeline.sinks.reducer.file import FileCleanupReducerSink
 from refiner.pipeline.sinks.reducer.lerobot import LeRobotMetaReduceSink
+from refiner.pipeline.sinks.reducer.zarr import (
+    ZarrCleanupReducerSink,
+    ZarrMergeReducerSink,
+)
 
 __all__ = [
     "FileCleanupReducerSink",
     "LeRobotMetaReduceSink",
+    "ZarrCleanupReducerSink",
+    "ZarrMergeReducerSink",
 ]
